working version (both random and from db)

+++ apl-api/src/main/java/com/apollocurrency/aplwallet/api/trading/SimpleTradingEntry.java
@@ -12,6 +12,7 @@ package com.apollocurrency.aplwallet.api.trading;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -23,12 +24,12 @@ import lombok.NoArgsConstructor;
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Schema(name = "SimpleTradingEntry", description = "Simple trading entry for graph data.")
-public class SimpleTradingEntry {
+public class SimpleTradingEntry {    
     public Integer time;
-    public double open;
-    public double close;
-    public double low; 
-    public double high;
-    public double volumefrom;
-    public double volumeto;    
+    public BigDecimal open;
+    public BigDecimal close;
+    public BigDecimal low; 
+    public BigDecimal high;
+    public BigDecimal volumefrom;
+    public BigDecimal volumeto;    
 }
